requirements.txt
jq==1.4.1
python-dotenv==1.0.0
requests==2.31.0
